feat: Keep telemetry context between microservices
To make it so that a context is sent between microservices (like when room service calls user service), we have to inject the context inside the http request that's about to be sent to the other microservice. It does make sending http requests a bit more verbose, and we could probably write some wrappers around it, but for once I'll keep things explicit.