fix: Don't push response size that's < 0 to Prometheus
This is a bit of an unusual case. The test itself sends a request with a missing ID, so it would ultimately return an error status code with no body. However, the "size" of the response would for some reason be -1, which our Prometheus client (the microservice) would try to add to the metrics. You're not allowed to decrease counters in Prometheus, so the whole thing would panic.